rename package to proxy
