+++ pkg/config/config.go
@@ -1,5 +1,4 @@
 package config
 
 type Config struct {
-
 }
