+++ pkg/proxy/buffered_read_conn.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package proxy
 
 import (
 	"bufio"
