+++ pkg/proxy/conn.go
@@ -1,4 +1,4 @@
-package server
+package proxy
 
 import (
 	"context"
@@ -14,8 +14,8 @@ import (
 const (
 	connStatusDispatching int32 = iota
 	connStatusReading
-	connStatusShutdown      // Closed by server.
-	connStatusWaitShutdown  // Notified by server to close.
+	connStatusShutdown     // Closed by server.
+	connStatusWaitShutdown // Notified by server to close.
 )
 
 type clientConn struct {
