+++ pkg/proxy/conn_handshake.go
@@ -1,4 +1,4 @@
-package server
+package proxy
 
 import (
 	"bytes"
