+++ pkg/proxy/conn_util.go
@@ -1,4 +1,4 @@
-package server
+package proxy
 
 import (
 	"context"
