+++ pkg/proxy/interface.go
@@ -1,4 +1,4 @@
-package server
+package proxy
 
 import (
 	"context"
