+++ pkg/proxy/packetio.go
@@ -32,7 +32,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package proxy
 
 import (
 	"bufio"
